requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
Flask-JWT-Extended==4.3.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
protobuf==3.19.3
PyJWT==2.3.0
python-dotenv==0.19.2
requests==2.26.0
soupsieve==2.3.1
urllib3==1.26.7
Werkzeug==2.0.2
